send results to participants
When the answer is revealed, the answers of all participants are
anonymized and then sent to the server and distributed.
This enables clients to also display a vote count or where other users
clicked in an image.

Also made the shared reveal function check the question type.
This obsoletes click.type